web-dj/requirements.txt

asgiref==3.4.1
async-generator==1.10

bcrypt==3.1.7

coreapi==2.3.3

decorator==5.0.9

Django==3.2.9
django-cors-headers==3.10.0
django-multi-email-field==0.6.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
django-rest-swagger==2.2.0
httplib2==0.14.0

Jinja2==2.10.1


mysqlclient==2.0.3
nbclient==0.5.9

numpy==1.19.5
oauthlib==3.1.0

passlib==1.7.4

Pillow==7.0.0

Pygments==2.11.1
PyJWT==1.7.1


pytz==2021.3

QtPy==2.0.0
requests==2.22.0

six==1.16.0
sqlparse==0.4.2


urllib3==1.25.11
virtualenv==20.10.0

zipp==3.6.0
